fix(SplitCarTest, SplitCar, Car): 해당 로직 CLASS model 디렉토리 하위 위치 이동
