+++ src/test/java/racingcar/SplitCarTest.java
@@ -2,6 +2,8 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.model.Car;
+import racingcar.model.SplitCar;
 
 import java.util.ArrayList;
